Refactor prompt structure
