+++ haystack/nodes/prompt/__init__.py
@@ -1,2 +1,5 @@
-from haystack.nodes.prompt.prompt_node import PromptNode, PromptTemplate, PromptModel, BaseOutputParser, AnswerParser
+from haystack.nodes.prompt.prompt_node import PromptNode
+from haystack.nodes.prompt.prompt_template import PromptTemplate
+from haystack.nodes.prompt.prompt_model import PromptModel
+from haystack.nodes.prompt.shapers import BaseOutputParser, AnswerParser
 from haystack.nodes.prompt.providers import PromptModelInvocationLayer
